Unit test CiteAttribution component
